+++ nuxt.config.ts
@@ -2,11 +2,6 @@ import process from 'node:process'
 
 export default defineNuxtConfig({
   srcDir: 'src',
-  runtimeConfig: {
-    public: {
-      websocketPort: 3030,
-    },
-  },
   app: {
     head: {
       link: [
@@ -139,8 +134,9 @@ export default defineNuxtConfig({
         base: './data',
       },
     },
-    // devServer: {
-    //   watch: ['./data'],
-    // },
+    experimental: {
+      websocket: true,
+      tasks: true,
+    },
   },
 })
