+++ src/plugins/settings.ts
@@ -1,7 +1,11 @@
 import type { CompleteConfig } from '~/types'
 
+/**
+ * @todo create a hot reboot configuration. Receive data from websocket
+ */
 export default defineNuxtPlugin(async () => {
   const { on } = useWebsocket()
+
   on('config:update', () => {
     reloadNuxtApp({
       force: true,
