+++ src/server/plugins/01.config-loader.ts
@@ -1,0 +1,6 @@
+/**
+ * @todo remove plugin when hooks appear in nitro task
+ */
+export default defineNitroPlugin(async () => {
+  runTask('config:update')
+})
