+++ src/server/tasks/config/update.ts
@@ -1,0 +1,9 @@
+export default defineTask({
+  meta: {
+    name: 'config:update',
+  },
+  async run() {
+    const config = await loadConfig()
+    await setConfig(config)
+  },
+})
